-x List Removed Project -x Restore Migration 	- move RemovedProject to project dir 	- intall run install
Note: MigrationHandler help migrate database files stored on Data/migrations

	-SchemaLoad Datase system
		if clDataSchema =>
			load data.schema.xml
			initDbSchemaBuilder - remove setting - automatically

x enum property must be start value, name=value

+++ src/Lib/igk/Lib/Classes/Database/SchemaBuilder/SchemaDiagramVisitor.php
@@ -6,7 +6,9 @@ namespace IGK\Database\SchemaBuilder;
 
 use Exception;
 use IGK\Controllers\BaseController;
+use IGK\Database\DbColumnInfo;
 use IGK\Database\DbSchemasConstants;
+use IGK\Database\SchemaBuilder\Entities\Factory;
 use IGK\Helper\Activator;
 use IGK\System\Console\Logger as ConsoleLogger;
 use IGK\System\Database\SchemaAddColumnMigration;
@@ -26,8 +28,9 @@ class SchemaDiagramVisitor extends DiagramVisitor
 {
     private $m_data;
     private $m_controller;
-    private $m_migrations = [];
-    private $operation = 'migrate';
+    private $m_migrations = []; 
+    private $m_entityHandler;
+    private $m_operation;
     /**
      * init SchemaDiagram
      * @param BaseController $controller 
@@ -38,7 +41,7 @@ class SchemaDiagramVisitor extends DiagramVisitor
     {
         $this->m_controller = $controller;
         $this->m_data = $schemaInfo;
-        $this->operation = $operation;
+        $this->m_operation = $operation;
     }
     /**
      * visit diagram 
@@ -91,39 +94,35 @@ class SchemaDiagramVisitor extends DiagramVisitor
             $index = $mig->$f();
             $index->setup($props['table'], $props['property']);
         } else {
-
-            switch ($migration->type) {
-                case 'dropEntity':
-                    $migrations = new SchemaBuilderMigration;
-                    $migrations->controller = $this->m_controller;
-                    $mig = new SchemaDeleteTableMigration($migrations);
-                    $mig->tables = [$migration->properties['name']];
-                    $r->migrations[] = $mig;
-                    break;
-                case 'addColumn':
-                    $mig = new SchemaBuilderMigration;
-                    $mig->controller = $this->m_controller;
-                    $cl = $mig->addColumn();
-                    $cl->setup(
-                        $props['table'],
-                        $props['property']->getColumnInfo(),
-                        igk_getv($props, 'after')
-                    );
-                    break;
-                case 'dropColumn':
-                    $mig = new SchemaBuilderMigration;
-                    $mig->controller = $this->m_controller;
-                    $cl = $mig->dropColumn();
-                    $cl->setup(
-                        $props['table'],
-                        $props['property']
-                    );
-                    break;
-                    // case 'changeColumn':
-                    //     break;
-                default:
-                    throw new Exception('type not supported');
-                    break;
+            if ($p = Factory::Create($migration->type)) { 
+                $mig = $p->setup($this->m_controller, $this->m_data, $props);
+                $this->m_entityHandler[] = $p;
+                //update schema
+                //$p->updateSchema($this->m_data);
+            } else { 
+                switch ($migration->type) {
+                    case 'dropEntity':
+                        $migrations = new SchemaBuilderMigration;
+                        $migrations->controller = $this->m_controller;
+                        $mig = new SchemaDeleteTableMigration($migrations);
+                        $mig->tables = [$migration->properties['name']];
+                        $r->migrations[] = $mig;
+                        break; 
+                    case 'dropColumn':
+                        $mig = new SchemaBuilderMigration;
+                        $mig->controller = $this->m_controller;
+                        $cl = $mig->dropColumn();
+                        $cl->setup(
+                            $props['table'],
+                            $props['property']
+                        );
+                        break;
+                        // case 'changeColumn':
+                        //     break;
+                    default:
+                        throw new Exception('type not supported');
+                        break;
+                }
             }
         }
         if ($mig) {
@@ -141,8 +140,10 @@ class SchemaDiagramVisitor extends DiagramVisitor
     {
         if ($this->m_migrations) {
             $this->callback = function ($e) {
+                // update database with migration - setting 
                 $ctrl = $e->args['ctrl'];
-                $type = $e->args['type'];
+                $type = $e->args['type']; 
+
                 $r = DbSchemasConstants::Downgrade;
                 if ($ctrl === $this->m_controller) {
                     foreach ($this->m_migrations as $mig) {
